Move `RenderEncoder` to `util`
